percussion logic

+++ include/manual.h
@@ -330,7 +330,7 @@ float remap(float v, float oldmin, float oldmax, float newmin, float newmax) {
 // drawbars[7]: 1 3/5' (15th)
 // drawbars[8]: 1 1/3' (19th)
 // drawbars[9]: 1' (22nd)
-uint32_t manual_fill_volumes(uint64_t keys, uint8_t drawbars[10], uint16_t ret[92]) {
+uint32_t manual_fill_volumes(uint64_t keys, uint16_t drawbars[10], uint16_t ret[92]) {
     float drawvols[] = {0, 1.414, 2, 2.828, 5, 5.657, 8, 11.31, 16};
 
     // The total possible gain per tonewheel, if all keys are down and
